refactoring and automatic scripts
