--- influx_si/bin/influx_i.py
@@ -1,1 +1,0 @@
-influx_s.py
